requirements.txt
pyscreenshot~=3.0
Pillow~=7.1.2
PySide2~=5.15.2.1
jsonpickle~=3.0.1
PyAutoGUI~=0.9.53